+++ streamlit_apps/neurosynchrony.py
@@ -18,7 +18,8 @@ from streamlit_apps.mert_components import (render_abnormalities,
                                             render_documents,
                                             render_eeg_review, render_notes,
                                             render_protocol_page,
-                                            get_report_addendum_eeg_id)
+                                            get_report_addendum_eeg_id,
+                                            render_eeg_history)
 from streamlit_dashboards import eeg_visualization_dashboard
 from streamlit_dashboards import ecg_visualization_dashboard
 from utils.helpers import calculate_age
@@ -283,33 +284,33 @@ if "tab" in st.session_state:
                             st.success(f"Neuroref Cz {report_id} successfully deleted!")
 
                 eegid = st.session_state["eegid"]
-                eeg_bucket = "lake-superior-prod"
-                eeg_s3_path = f"bronze/eegs/clinical/{eegid}.dat"
-                if eegid:
-                  if not key_exists(eeg_bucket, eeg_s3_path):
-                      edf_path = f"bronze/eegs/clinical/{eegid}.edf"
-                      if not key_exists(eeg_bucket, edf_path):
-                          raise Exception("EEG could not be found.")
-                      else:
-                          eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=edf_path)
-                          eeg_content = eeg_obj["Body"].read()
-                          fname = f"{eegid}.edf"
-                  else:
-                      eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=eeg_s3_path)
-                      eeg_content = eeg_obj["Body"].read()
-                      fname = f"{eegid}.dat"
-                else:
-                    eeg_content = "empty file"
-                    fname = "empty_file.txt"
-                if st.download_button(label="Download EEG", data=eeg_content, file_name=fname):
-                  try:
-                      st.write(f"EEG:'{eegid}' downloaded successfully")
-                  except NoCredentialsError:
-                      st.error("Error: Unable to locate credentials")
-                  except PartialCredentialsError:
-                      st.error("Error: Incomplete credentials provided")
-                  except Exception as e:
-                      st.error(f"Error: {e}")
+                # eeg_bucket = "lake-superior-prod"
+                # eeg_s3_path = f"bronze/eegs/clinical/{eegid}.dat"
+                # if eegid:
+                #   if not key_exists(eeg_bucket, eeg_s3_path):
+                #       edf_path = f"bronze/eegs/clinical/{eegid}.edf"
+                #       if not key_exists(eeg_bucket, edf_path):
+                #           raise Exception("EEG could not be found.")
+                #       else:
+                #           eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=edf_path)
+                #           eeg_content = eeg_obj["Body"].read()
+                #           fname = f"{eegid}.edf"
+                #   else:
+                #       eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=eeg_s3_path)
+                #       eeg_content = eeg_obj["Body"].read()
+                #       fname = f"{eegid}.dat"
+                # else:
+                #     eeg_content = "empty file"
+                #     fname = "empty_file.txt"
+                # if st.download_button(label="Download EEG", data=eeg_content, file_name=fname):
+                #   try:
+                #       st.write(f"EEG:'{eegid}' downloaded successfully")
+                #   except NoCredentialsError:
+                #       st.error("Error: Unable to locate credentials")
+                #   except PartialCredentialsError:
+                #       st.error("Error: Incomplete credentials provided")
+                #   except Exception as e:
+                #       st.error(f"Error: {e}")
 
                 st.header("EEG History")
                 with st.form("data_editor_form", border=False):
@@ -544,33 +545,33 @@ else:
                         st.success(f"Neuroref Cz {report_id} successfully deleted!")
 
             eegid = st.session_state["eegid"]
-            eeg_bucket = "lake-superior-prod"
-            eeg_s3_path = f"bronze/eegs/clinical/{eegid}.dat"
-            if eegid:
-              if not key_exists(eeg_bucket, eeg_s3_path):
-                  edf_path = f"bronze/eegs/clinical/{eegid}.edf"
-                  if not key_exists(eeg_bucket, edf_path):
-                      raise Exception("EEG could not be found.")
-                  else:
-                      eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=edf_path)
-                      eeg_content = eeg_obj["Body"].read()
-                      fname = f"{eegid}.edf"
-              else:
-                  eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=eeg_s3_path)
-                  eeg_content = eeg_obj["Body"].read()
-                  fname = f"{eegid}.dat"
-            else:
-                eeg_content = "empty file"
-                fname = "empty_file.txt"
-            if st.download_button(label="Download EEG", data=eeg_content, file_name=fname):
-              try:
-                  st.write(f"EEG:'{eegid}' downloaded successfully")
-              except NoCredentialsError:
-                  st.error("Error: Unable to locate credentials")
-              except PartialCredentialsError:
-                  st.error("Error: Incomplete credentials provided")
-              except Exception as e:
-                  st.error(f"Error: {e}")
+            # eeg_bucket = "lake-superior-prod"
+            # eeg_s3_path = f"bronze/eegs/clinical/{eegid}.dat"
+            # if eegid:
+            #   if not key_exists(eeg_bucket, eeg_s3_path):
+            #       edf_path = f"bronze/eegs/clinical/{eegid}.edf"
+            #       if not key_exists(eeg_bucket, edf_path):
+            #           raise Exception("EEG could not be found.")
+            #       else:
+            #           eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=edf_path)
+            #           eeg_content = eeg_obj["Body"].read()
+            #           fname = f"{eegid}.edf"
+            #   else:
+            #       eeg_obj = s3.get_object(Bucket=eeg_bucket, Key=eeg_s3_path)
+            #       eeg_content = eeg_obj["Body"].read()
+            #       fname = f"{eegid}.dat"
+            # else:
+            #     eeg_content = "empty file"
+            #     fname = "empty_file.txt"
+            # if st.download_button(label="Download EEG", data=eeg_content, file_name=fname):
+            #   try:
+            #       st.write(f"EEG:'{eegid}' downloaded successfully")
+            #   except NoCredentialsError:
+            #       st.error("Error: Unable to locate credentials")
+            #   except PartialCredentialsError:
+            #       st.error("Error: Incomplete credentials provided")
+            #   except Exception as e:
+            #       st.error(f"Error: {e}")
 
             st.header("EEG History")
             with st.form("data_editor_form", border=False):
@@ -614,6 +615,10 @@ else:
             st.divider()
 
             render_abnormalities(data_manager)
+
+            st.divider()
+
+            render_eeg_history(data_manager)
 
 
     with tab2: